+++ EveryTime/Controller/MainView/MainViewController+Scroll.swift
@@ -9,33 +9,33 @@
 import UIKit
 
 extension MainViewController: UIScrollViewDelegate {
-    func scrollViewWillBeginDecelerating(_ scrollView: UIScrollView) {
-        executeState(state: .Hide)
-    }
-    func scrollViewDidEndDecelerating(_ scrollView: UIScrollView) {
-        executeState(state: .Show)
-    }
+//    func scrollViewWillBeginDecelerating(_ scrollView: UIScrollView) {
+//        executeState(state: .Hide)
+//    }
+//    func scrollViewDidEndDecelerating(_ scrollView: UIScrollView) {
+//        executeState(state: .Show)
+//    }
+//    
+//    func executeState(state: ScrollingState) {
+//        switch state {
+//        case .Show:
+//            showStepButtonAnimation()
+//        case .Hide:
+//            hideStepButtonAnimation()
+//        case .Idle:
+//            break
+//        }
+//    }
     
-    func executeState(state: ScrollingState) {
-        switch state {
-        case .Show:
-            showStepButtonAnimation()
-        case .Hide:
-            hideStepButtonAnimation()
-        case .Idle:
-            break
-        }
-    }
-    
-    func hideStepButtonAnimation() {
-        UIView.animate(withDuration: 0.15, delay: 0.0, options: [.curveEaseInOut], animations: {
-            self.mainViewView.addRecipeButton.center.y = self.view.frame.maxY + 50.0
-        }, completion: nil)
-    }
-    
-    func showStepButtonAnimation() {
-        UIView.animate(withDuration: 0.2, delay: 0.2, options: [.curveEaseInOut], animations: {
-            self.mainViewView.addRecipeButton.center.y = self.view.frame.maxY - 50.0
-        }, completion: nil)
-    }
+//    func hideStepButtonAnimation() {
+//        UIView.animate(withDuration: 0.15, delay: 0.0, options: [.curveEaseInOut], animations: {
+//            self.mainViewView.addRecipeButton.center.y = self.view.frame.maxY + 50.0
+//        }, completion: nil)
+//    }
+//
+//    func showStepButtonAnimation() {
+//        UIView.animate(withDuration: 0.2, delay: 0.2, options: [.curveEaseInOut], animations: {
+//            self.mainViewView.addRecipeButton.center.y = self.view.frame.maxY - 50.0
+//        }, completion: nil)
+//    }
 }