+++ EveryTime/View/AddRecipe/AddRecipeBaseCell.swift
@@ -12,7 +12,7 @@ class AddRecipeBaseCell: UICollectionViewCell {
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-        self.setupView()
+        self.backgroundColor = UIColor.clear
     }
     
     required init?(coder aDecoder: NSCoder) {