+++ WhatsNew.md
@@ -1,5 +1,16 @@
 #  Whats New Description for App Store
 
+## Version 1.1.12
+Adding Recipe/Steps
+    - Textfields now have a default value
+Main View    
+    - The current working step now appears on the main view
+    - Adding a recipe now moved to the top right navigation bar and removed from the bottom
+    - Clearing all recipes is now located within Settings
+Other
+- Typo in settings caption
+
+
 ## Version 1.1.11
 Taptic Feedback
     - Now on pause buttons